+++ api/V0KillTask.go
@@ -11,7 +11,7 @@ import (
 
 // V0KillTask will kill the given task id
 // example:
-// curl -X GET 127.0.0.1:10000/v0/task/kill/{count of instances} -d 'JSON'
+// curl -X GET http://user:password@127.0.0.1:10000/v0/task/kill/{task id} -d 'JSON'
 func V0KillTask(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	auth := CheckAuth(r, w)