+++ api/V0ReflateAgent.go
@@ -9,9 +9,9 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// V0ReflateMissingK3SAgent will scale the zookeeper service
+// V0ReflateMissingK3SAgent will scale the agent service
 // example:
-// curl -X GET 127.0.0.1:10000/v0/zookeeper/reflate -d 'JSON'
+// curl -X GET 127.0.0.1:10000/v0/agent/reflate -d 'JSON'
 func V0ReflateK3SAgent(w http.ResponseWriter, r *http.Request) {
 	logrus.Debug("HTTP GET V0RestartMissingK3SAgent")
 	auth := CheckAuth(r, w)