+++ api/api.go
@@ -23,10 +23,10 @@ func SetConfig(cfg *cfg.Config) {
 func Commands() *mux.Router {
 	rtr := mux.NewRouter()
 	rtr.HandleFunc("/v0/container/{taskID}", V0StatusContainer).Methods("GET")
-	rtr.HandleFunc("/v0/zookeeper/scale/{count}", V0ScaleK3SAgent).Methods("GET")
-	rtr.HandleFunc("/v0/zookeeper/reflate", V0ReflateK3SAgent).Methods("GET")
-	rtr.HandleFunc("/v0/k3s/scale/{count}", V0ScaleK3S).Methods("GET")
-	rtr.HandleFunc("/v0/k3s/reflate", V0ReflateK3S).Methods("GET")
+	rtr.HandleFunc("/v0/agent/scale/{count}", V0ScaleK3SAgent).Methods("GET")
+	rtr.HandleFunc("/v0/agent/reflate", V0ReflateK3SAgent).Methods("GET")
+	rtr.HandleFunc("/v0/server/scale/{count}", V0ScaleK3S).Methods("GET")
+	rtr.HandleFunc("/v0/server/reflate", V0ReflateK3S).Methods("GET")
 	rtr.HandleFunc("/v0/task/kill/{id}", V0KillTask).Methods("GET")
 
 	return rtr